+++ src/BabelFish/DataModel/Definitions/ShowWhenBase.cs
@@ -1,16 +1,15 @@
-﻿using Scopos.BabelFish.Converters;
-using Scopos.BabelFish.DataModel.Definitions;
-using System;
-using System.Collections.Generic;
 using System.ComponentModel;
-using System.Text;
 
 namespace Scopos.BabelFish.DataModel.Definitions {
 
-
+    /// <summary>
+    /// ShowWhen operations describe logic for when a <seealso cref="ResultListFormat">RESULT LIST FORMAT</seealso>
+    /// <seealso cref="ResultListDisplayColumn"/>, <seealso cref="ClassSet"/>, or SpanningText is included and displayed.
+    /// <para>ShowWhenBase is the abstract base class for describing this logic.</para>
+    /// </summary>
     [Serializable]
     [G_NS.JsonConverter( typeof( G_BF_NS_CONV.ShowWhenBaseConverter ) )]
-    public abstract class ShowWhenBase: IReconfigurableRulebookObject {
+    public abstract class ShowWhenBase : IReconfigurableRulebookObject {
 
         /// <summary>
         /// Concret class identifier.